Fix issues caught by other compilers

+++ src/finite_element/export_finite_element.cpp
@@ -492,14 +492,14 @@ private:
 	void clearTemplates()
 	{
 		const size_t ntCount = this->nodeTemplates.size();
-		for (int nt = 0; nt < ntCount; ++nt)
+		for (size_t nt = 0; nt < ntCount; ++nt)
 		{
 			delete this->nodeTemplates[nt];
 		}
 		this->nodeTemplates.clear();
 		this->nodeTemplate = nullptr;
 		const size_t etCount = this->elementTemplates.size();
-		for (int et = 0; et < etCount; ++et)
+		for (size_t et = 0; et < etCount; ++et)
 		{
 			delete this->elementTemplates[et];
 		}
